+++ src/pages/process/components/processchart/index.js
@@ -13,9 +13,9 @@ import UpdateProcess from "./components/updateprocess";
 import { FcProcess } from "react-icons/fc";
 
 function ProcessChart() {
-  const API_BASE_URL = "http://localhost:3003";
+  const API_BASE_URL = "http://localhost:3005";
   const [tasks, setTasks] = useState([]);
-  const [view, setView] = useState(ViewMode.Hour);
+  const [view, setView] = useState(ViewMode.Day);
   const [isChecked, setIsChecked] = useState(true);
   const [selectedTaskData, setSelectedTaskData] = useState(null);
   const [open, setOpen] = useState(false);
@@ -23,6 +23,7 @@ function ProcessChart() {
   const [showUpdate, setShowUpdate] = useState(false);
   const [updateModalVisible, setUpdateModalVisible] = useState(false);
   const [selectedProcessData, setSelectedProcessData] = useState(null);
+  const [loading, setLoading] = useState(true);
 
   const handleEditProcessClick = (selectedTaskData) => {
     setSelectedProcessData(selectedTaskData);
@@ -41,6 +42,22 @@ function ProcessChart() {
   useEffect(() => {
     fetchProcessData();
   }, []);
+
+  function formatDuration(start, end) {
+    const durationInmilliseconds = end - start;
+    const hours = Math.floor(durationInmilliseconds / (1000 * 60 * 60));
+    const minutes = Math.floor(
+      (durationInmilliseconds % (1000 * 60 * 60)) / (1000 * 60)
+    );
+
+    if (hours === 0) {
+      return `${minutes} minute(s)`;
+    } else if (minutes === 0) {
+      return `${hours} hour(s)`;
+    } else {
+      return `${hours} hour(s) and ${minutes} minute(s)`;
+    }
+  }
 
   const menu = (
     <Menu>
@@ -64,8 +81,8 @@ function ProcessChart() {
   const fetchProcessData = async () => {
     try {
       const [response, subprocessResponse] = await Promise.all([
-        axios.get("http://localhost:3003/process"),
-        axios.get("http://localhost:3003/subprocess"),
+        axios.get("http://localhost:3005/process"),
+        axios.get("http://localhost:3005/subprocess"),
       ]);
 
       const process = response.data.data;
@@ -88,28 +105,12 @@ function ProcessChart() {
       });
 
       let count = 1; // Initialize a count variable
-      const mappedProcesses = [];
+      const mappedProcesses = {};
 
       // Create a map of subprocesses by their _id for efficient lookup
       const subprocessMap = {};
       // Iterate through the process array
       process.forEach((item) => {
-        function formatDuration(start, end) {
-          const durationInMilliseconds = end - start;
-          const hours = Math.floor(durationInMilliseconds / (1000 * 60 * 60));
-          const minutes = Math.floor(
-            (durationInMilliseconds % (1000 * 60 * 60)) / (1000 * 60)
-          );
-
-          if (hours === 0) {
-            return `${minutes} minute(s)`;
-          } else if (minutes === 0) {
-            return `${hours} hour(s)`;
-          } else {
-            return `${hours} hour(s) and ${minutes} minute(s)`;
-          }
-        }
-
         const ProcessdurationInHours = formatDuration(
           new Date(item.start),
           new Date(item.end)
@@ -118,53 +119,96 @@ function ProcessChart() {
           "Formatted ProcessdurationInHours: " + ProcessdurationInHours
         );
 
-        // Map the item to the desired format
-        const mappedItem = {
-          key: item._id,
-          start: new Date(item.start),
-          end: new Date(item.end),
-          name: item.name,
-          id: item.name,
-          processId: item.pid,
-          humanresource: item.humanResource,
-          rawmaterial: item.rawMaterial,
-          duration: ProcessdurationInHours,
-          type: "project",
-          displayOrder: count++,
-        };
-        mappedProcesses.push(mappedItem);
+        // Check if the item is already in mappedProcesses to avoid duplicates
+        if (!mappedProcesses[item._id]) {
+          const mappedItem = {
+            key: item._id,
+            start: new Date(item.start),
+            end: new Date(item.end),
+            name: item.name,
+            id: item.name,
+            processId: item.pid,
+            humanresource: item.humanResource,
+            rawmaterial: item.rawMaterial,
+            duration: ProcessdurationInHours,
+            type: "project",
+            displayOrder: count++,
+            hideChildren: false,
+          };
+          mappedProcesses[item._id] = mappedItem;
 
-        if (item.subprocesses.length > 0) {
-          let lastItem = [];
-          item.subprocesses.forEach((subitem) => {
-            const subdurationInHours = formatDuration(
-              new Date(subitem.substart),
-              new Date(subitem.subend)
-            );
-            const mappedItem = {
-              key: subitem._id,
-              start: new Date(subitem.substart),
-              end: new Date(subitem.subend),
-              name: subitem.subname,
-              id: subitem.subname,
-              subhumanresource: subitem.humanResource,
-              subrawmaterial: subitem.rawMaterial,
-              duration: subdurationInHours,
-              type: "task",
-              project: item.name,
-              displayOrder: count++,
-              dependencies: lastItem,
-            };
-            mappedProcesses.push(mappedItem);
-            lastItem = [subitem.subname];
-          });
+          if (item.subprocesses) {
+            let lastItem = null;
+            item.subprocesses.forEach((subitem, index) => {
+              const subdurationInHours = formatDuration(
+                new Date(subitem.substart),
+                new Date(subitem.subend)
+              );
+              let subMappedItem = {
+                key: subitem._id,
+                start: new Date(subitem.substart),
+                end: new Date(subitem.subend),
+                name: subitem.subname,
+                id: subitem.subname,
+                subprocessId: subitem.subId,
+                subhumanresource: subitem.humanResource,
+                subrawmaterial: subitem.rawMaterial,
+                duration: subdurationInHours,
+                type: "task",
+                project: item.name,
+                displayOrder: count++,
+              };
+
+              if (index) {
+                subMappedItem = { ...subMappedItem, dependencies: [lastItem] };
+              }
+
+              lastItem = subitem.subname;
+              mappedProcesses[subitem._id] = subMappedItem;
+            });
+          }
         }
       });
 
-      // Log the mapped processes
-      console.log(mappedProcesses);
+      // Convert the mapped processes map to an array
+      const processesArray = Object.values(mappedProcesses);
 
-      setTasks(mappedProcesses);
+      // Log the mapped processes
+      console.log(processesArray, "sbdjhsbf");
+
+      setTasks(
+        processesArray.map((task) => ({
+          ...task,
+          name: (
+            <>
+              <FcProcess />{" "}
+              {task.processId ? (
+                <span style={{ color: "green" }}>
+                  {task.processId.toUpperCase()}
+                </span>
+              ) : (
+                ""
+              )}{" "}
+              {task.subprocessId ? (
+                <span style={{ color: "blue" }}>
+                  {task.subprocessId.toUpperCase()}
+                </span>
+              ) : (
+                ""
+              )}{" "}
+              {task.type === "project"
+                ? task.name
+                : task.subprocesses
+                ? `${task.subprocesses.subname} - ${task.name}`
+                : task.name}
+              <ProcessTags
+                status={getStatusForProcess(task)}
+                style={{ display: "flex" }}
+              />
+            </>
+          ),
+        }))
+      );
     } catch (error) {
       console.error("Error fetching data:", error);
     }
@@ -184,46 +228,52 @@ function ProcessChart() {
     let newTasks = tasks.map((t) => (t.id === task.id ? task : t));
     if (task.project) {
       const [start, end] = getStartEndDateForProject(newTasks, task.project);
-      const project =
-        newTasks[newTasks.findIndex((t) => t.id === task.project)];
+      const projectIndex = newTasks.findIndex((t) => t.id === task.project);
+      const project = newTasks[projectIndex];
       if (
         project.start.getTime() !== start.getTime() ||
         project.end.getTime() !== end.getTime()
       ) {
-        const changedProject = { ...project, start, end };
+        const changedProject = {
+          ...project,
+          start,
+          end,
+        };
         newTasks = newTasks.map((t) =>
           t.id === task.project ? changedProject : t
         );
       }
     }
     setTasks(newTasks);
+    console.log("up", newTasks);
   };
 
-  const handleDblClick = (taskId) => {
-    console.log("Clicked task ID:", taskId);
+  // const handleDblClick = (taskId) => {
+  //   console.log("Clicked task ID:", taskId);
 
-    const selectedTaskData = tasks.find((task) => task.id === taskId._id);
-    console.log("Selected task data:", selectedTaskData);
+  //   const selectedTaskData = tasks.find((task) => task.id === taskId._id);
+  //   console.log("Selected task data:", selectedTaskData);
 
-    if (selectedTaskData) {
-      setSelectedTaskData(selectedTaskData);
-      // setIsSidebarOpen(true); // Open the sidebar
-    }
-  };
+  //   if (selectedTaskData) {
+  //     setSelectedTaskData(selectedTaskData);
+  //     // setIsSidebarOpen(true); // Open the sidebar
+  //   }
+  // };
 
   const handleClick = (task) => {
-    console.log("On Click event Id:" + task.id);
-    console.log("Clicked task ID:", task);
+    console.log("up-onc", task);
+    // console.log("On Click event Id:" + task.id);
+    // console.log("Clicked task ID:", task);
 
-    const selectedTaskData = tasks.find((taskid) => taskid.id === task.id);
-    console.log("Selected task data:", selectedTaskData);
+    // const selectedTaskData = tasks.find((taskid) => taskid.id === task.id);
+    // console.log("Selected task data:", selectedTaskData);
 
-    if (selectedTaskData) {
-      setSelectedTaskData(selectedTaskData);
-    }
+    // if (selectedTaskData) {
+    // }
+    setSelectedTaskData(task);
     setOpen(true);
 
-    console.log();
+    // console.log();
   };
 
   const handleMenuClick = (e) => {
@@ -243,14 +293,20 @@ function ProcessChart() {
   const handleDeleteProcess = async (prId, sprId) => {
     try {
       await axios.delete(`${API_BASE_URL}/process/${prId}`);
-      await axios.delete(`${API_BASE_URL}/subprocess/${sprId}`);
+      if (sprId) {
+        await axios.delete(`${API_BASE_URL}/subprocess/${sprId}`);
+      }
       alert("Process deleted successfully");
-
-      // Reload the current route
       window.location.reload();
     } catch (error) {
-      alert("Process data deleted");
+      console.error("Error deleting process:", error);
+      alert("Error deleting process");
     }
+  };
+
+  const handleExpanderClick = (task) => {
+    setTasks(tasks.map((t) => (t.id === task.id ? task : t)));
+    console.log("On expander click Id:" + task.id);
   };
 
   return (
@@ -260,7 +316,6 @@ function ProcessChart() {
           onViewModeChange={(viewMode) => setView(viewMode)}
           onViewListChange={setIsChecked}
           isChecked={isChecked}
-          defaultViewMode={ViewMode.Hour}
         />
       </Wrapper>
 
@@ -269,25 +324,14 @@ function ProcessChart() {
       {tasks.length > 0 ? (
         <Chart>
           <Gantt
-            tasks={tasks.map((task) => ({
-              ...task,
-              name: (
-                <>
-                  <FcProcess />{" "}
-                  {task.processId ? task.processId.toUpperCase() : ""}{" "}
-                  {task.name}{" "}
-                  <ProcessTags
-                    status={getStatusForProcess(task)}
-                    style={{ display: "flex" }}
-                  />
-                </>
-              ),
-            }))}
+            tasks={tasks}
             viewMode={view}
             onDateChange={handleTaskChange}
-            onDoubleClick={handleDblClick}
-            onClick={handleClick}
+            // onDoubleClick={handleDblClick}
+            onDoubleClick={handleClick}
             listCellWidth={isChecked ? "155px" : ""}
+            columnWidth={columnWidth}
+            onExpanderClick={handleExpanderClick}
           />
           <Wrapper
             style={{
